BancoAgil/requirements.txt
dateparser==1.2.1
Flask==3.1.1
flask_cors==6.0.1
pandas==1.5.3
protobuf==6.31.1
python-dotenv==1.1.0
tavily_python==0.3.3
Werkzeug==3.1.3
